bothelper/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
cryptography==2.7
Django==2.2.2
djangorestframework==3.9.4
future==0.17.1
Pillow==6.0.0
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
